* revert the latest patch from screamingwithnosound   NOTE: This did not work as intended

+++ src/game/game.h
@@ -207,9 +207,9 @@ typedef struct {
 	int (*GetModelAndName) (char *team, char *path, char *body, char *head, char *name);
 
 	/* managed memory allocation */
-	void *(*TagMalloc) (size_t size, uint16_t tag);
+	void *(*TagMalloc) (size_t size, int tag);
 	void (*TagFree) (void *block);
-	void (*FreeTags) (uint16_t tag);
+	void (*FreeTags) (int tag);
 
 	/* console variable interaction */
 	cvar_t *(*cvar) (char *var_name, char *value, int flags);
